增加一个二维数组表示棋盘的空结构
